pyproject.toml
[project]
name = "agent-db"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
pdf = ["pymupdf>=1.24"]
llm = ["anthropic>=0.40"]
all = ["pymupdf>=1.24", "anthropic>=0.40"]

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
